Scale Crop Coors for Different Image Sizes
To keep crops in the same relative positions even when the template
image and image being cropped are different sizes, scale the crop
coordinates based on the scaling done to display image and the size
ratios (for x and y dimensions) between template and to_crop images.
This helps keep the crop in the same anatomical position across images
of animals.