Add Saving and Loading Coordinate Files
After selecting a region, the user can now choose to save the
coordinates of the region to a machine-readable configuration file (INI
file). This file can then be loaded by the same or other instances of
the program to re-create the same region on the same or different image.
Config files are saved to and laoded from files the user chooses via
standard system dialogs. Error handling is present but untested.